chore: use right motiontype
