+++ crates/types/src/lib.rs
@@ -39,7 +39,7 @@ pub mod message_event;
 pub mod messages;
 pub mod motion_command;
 pub mod motion_runtime;
-pub mod motion_selection;
+pub mod motion_type;
 pub mod motor_commands;
 pub mod multivariate_normal_distribution;
 pub mod object_detection;
