+++ crates/world_state/src/behavior/switch.rs
@@ -1,12 +1,14 @@
 use std::time::Duration;
 
-use types::behavior_tree::MotionType;
+use types::motion_type::MotionType;
 
-use crate::behavior::node::{Blackboard};
+use crate::behavior::node::Blackboard;
 
 pub fn is_allowed_to_switch(blackboard: &mut Blackboard) -> bool {
     let parameters = &blackboard.parameters.allow_switch;
-    let time_since_last_switch = blackboard.world_state.now
+    let time_since_last_switch = blackboard
+        .world_state
+        .now
         .duration_since(blackboard.last_motion_switch_time)
         .unwrap_or(Duration::ZERO);
